Remove tl_macros; allow message acks.
Part of the implementation of all of this was to make WriteType object-safe so
Box<TLObject> can be serialized.

The codegen is much easier to read/write/debug than the macros, especially when
there are errors coming from within the generated code. So, there's now a
manual.tl file for the things that, for some reason, were commented out of the
official schema.tl.